requirements.txt
flask==3.1.0
flask-sqlalchemy==3.1.1
mysql-connector-python==9.2.0
python-dotenv==1.0.1
Werkzeug==3.1.3
gunicorn==21.2.0
torch==2.6.0
torchvision==0.21.0
Pillow==11.1.0
numpy==2.2.3
PyJWT==2.8.0 